+++ common/src/core/resource/layout_resource.cpp
@@ -245,12 +245,12 @@ QHash<int, QVariant> QnLayoutResource::data() const {
 /********* Properties getters and setters **********/
 
 /********* Cell aspect ratio propert **********/
-qreal QnLayoutResource::cellAspectRatio() const {
+float QnLayoutResource::cellAspectRatio() const {
     QMutexLocker locker(&m_mutex);
     return m_cellAspectRatio;
 }
 
-void QnLayoutResource::setCellAspectRatio(qreal cellAspectRatio) {
+void QnLayoutResource::setCellAspectRatio(float cellAspectRatio) {
     {
         QMutexLocker locker(&m_mutex);
         if(qFuzzyCompare(m_cellAspectRatio, cellAspectRatio))
